requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
python-dateutil==2.9.0.post0
scikit-learn==1.5.1
